requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyobjc-framework-Quartz>=9.2  # Required for keyboard simulation on macOS
